+++ src/Main/ExampleMaker.java
@@ -18,31 +18,31 @@ public class ExampleMaker {
 
     Random rand = new Random();
 
-    public float calcAngle(PointEx p1, PointEx p2, PointEx p3) {
+    public double calcAngle(PointEx p1, PointEx p2, PointEx p3) {
         // p1 jest wierzcholkiem, p1 is vertex
-        float p1_2 = calcLength(p1, p2);
-        float p1_3 = calcLength(p1, p3);
-        float p2_3 = calcLength(p2, p3);
+        double p1_2 = calcLength(p1, p2);
+        double p1_3 = calcLength(p1, p3);
+        double p2_3 = calcLength(p2, p3);
 
-        return (float) Math.acos((p1_2 * p1_2 + p1_3 * p1_3 - p2_3 * p2_3) / (2 * p1_2 * p1_3));
+        return  Math.acos((p1_2 * p1_2 + p1_3 * p1_3 - p2_3 * p2_3) / (2 * p1_2 * p1_3));
     }
 
-    float calcLength(PointEx from, PointEx to) {
-        return (float) Math.sqrt(((from.x - to.x) * (from.x - to.x)) + ((from.y - to.y) * (from.y - to.y)));
+    double calcLength(PointEx from, PointEx to) {
+        return  Math.sqrt(((from.x - to.x) * (from.x - to.x)) + ((from.y - to.y) * (from.y - to.y)));
     }
 
     static public List<Example> makeExamples() {
         List<Example> examplesList = new ArrayList<>();
         for (int i = 0; i < 60; i++) {
             double angle1 = Math.random() * Math.PI;
-            float tmpX = (float) Math.cos(angle1) * MyRobot.radius + MyRobot.p1.x;
-            float tmpY = (float) Math.sin(angle1) * MyRobot.radius + MyRobot.p1.y;
+            double tmpX =  Math.cos(angle1) * MyRobot.radius + MyRobot.p1.x;
+            double tmpY =  Math.sin(angle1) * MyRobot.radius + MyRobot.p1.y;
             for (int j = 0; j<30;j++) {
                 double angle2 = Math.random() * Math.PI * 2;
-                float x = (float) Math.cos(angle2) * MyRobot.radius + tmpX;
-                float y = (float) Math.sin(angle2) * MyRobot.radius + tmpY;
+                double x =  Math.cos(angle2) * MyRobot.radius + tmpX;
+                double y =  Math.sin(angle2) * MyRobot.radius + tmpY;
                 PointEx p = new PointEx(x, y);
-                examplesList.add(new Example(p, (float)angle1, (float)angle2));
+                examplesList.add(new Example(p, angle1, angle2));
             }
         }
