+++ src/Main/MyCircle.java
@@ -14,21 +14,21 @@ import processing.core.PApplet;
 public class MyCircle {
 
     PApplet aplet;
-    float x;
-    float y;
+    double x;
+    double y;
     int radius;
     int color;
 
-    public float distance(MyCircle otherPoint){
+    public double distance(MyCircle otherPoint){
         return ((this.x - otherPoint.x)*(this.x - otherPoint.x)) + ((this.y - otherPoint.y)*(this.y - otherPoint.y));
     }
     
-    public void update(float alpha, float G, MyCircle otherPoint) {
+    public void update(double alpha, double G, MyCircle otherPoint) {
         this.x += alpha * G * (otherPoint.x - this.x);
         this.y += alpha * G * (otherPoint.y - this.y);
     }
     
-    public MyCircle(PApplet aplet, float x, float y, int radius, int color) {
+    public MyCircle(PApplet aplet, double x, double y, int radius, int color) {
         this.aplet = aplet;
         this.x = x;
         this.y = y;
@@ -36,7 +36,7 @@ public class MyCircle {
         this.color = color;
     }
 
-    public MyCircle(float x, float y, int radius, int color) {
+    public MyCircle(double x, double y, int radius, int color) {
         this.x = x;
         this.y = y;
         this.radius = radius;
@@ -46,7 +46,7 @@ public class MyCircle {
     public void display() {
         aplet.stroke(255);
         aplet.fill(204, 102, 0);
-        aplet.ellipse(x, y, radius, radius);
+        aplet.ellipse((float)x, (float)y, radius, radius);
         aplet.fill(100, 100, 255);
     }
